github_issue_lister_cli/pyproject.toml
[project]
name = "github-issue-lister-cli"
version = "0.1.0"
description = "A CLI tool to list GitHub issues."
dependencies = [
    "PyGithub==2.6.1",
    "typer==0.16.0"
]

[project.scripts]
github-lister = "mcp-typer:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
